generate pages dynamically based on new tags
REFACTOR
we go to all posts and fetch their tags array and create a SET of it and then you use that array set to create dynamic paths to each tag.

Remember
A getStaticPaths function should always return a list of objects containing params (what to call each page route) and optionally any props (data that you want to pass to those pages). Earlier, you defined each tag name that you knew was used in your blog and passed the entire list of posts as props to each page.